style(): rename all references of tone to note
Tone was a wrong naming for this, because it is a wave with certain
frequency. it was actually a number representing a certain note, with
progression from 0 upwards and with each step a change of half a tone.

Because of that mistake every wrong occurrence of tone is
changed to note.